sgminers as external deps
