+++ NiceHashMiner.sln
@@ -114,6 +114,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCMinerMTP", "src\Miners\CC
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CCMinerTpruvot", "src\Miners\CCMinerTpruvot\CCMinerTpruvot.csproj", "{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SgminerAvemore", "src\Miners\SgminerAvemore\SgminerAvemore.csproj", "{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SgminerGM", "src\Miners\SgminerGM\SgminerGM.csproj", "{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -894,6 +898,38 @@ Global
 		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|Win32.Build.0 = Release|Any CPU
 		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|x64.ActiveCfg = Release|Any CPU
 		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB}.Release|x64.Build.0 = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|Win32.Build.0 = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Debug|x64.Build.0 = Debug|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|Win32.ActiveCfg = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|Win32.Build.0 = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|x64.ActiveCfg = Release|Any CPU
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624}.Release|x64.Build.0 = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|Win32.ActiveCfg = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|Win32.Build.0 = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Debug|x64.Build.0 = Debug|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|Win32.ActiveCfg = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|Win32.Build.0 = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|x64.ActiveCfg = Release|Any CPU
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1}.Release|x64.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -942,6 +978,8 @@ Global
 		{138E97ED-8C5B-4C31-90F8-4260714A56E7} = {BB931C9A-EE48-42A6-B5F3-3A193B1A051A}
 		{C6E4069B-2305-4CDA-9F4E-0B79A8599935} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 		{3651E4F1-76B9-4634-A8D2-2A05E60C16CB} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
+		{F8FC33DC-2F03-4BA3-BDF7-D72CA673E624} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
+		{DF3E6CDA-C809-4C2D-BF95-68E4293E6EC1} = {F40126B4-5B3B-4085-B47D-A5EF4B51570C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6E8897DC-99C6-4585-A2BA-CB194F3B1AB7}
