+++ src/Miners/SgminerGM/PluginInternalSettings.cs
@@ -1,0 +1,24 @@
+﻿using MinerPluginToolkitV1.Configs;
+using MinerPluginToolkitV1.ExtraLaunchParameters;
+using MinerPluginToolkitV1.SgminerCommon;
+using System.Collections.Generic;
+
+namespace SgminerGM
+{
+    internal static class PluginInternalSettings
+    {
+        internal static MinerOptionsPackage MinerOptionsPackage = SgminerOptionsPackage.DefaultMinerOptionsPackage;
+
+        internal static MinerSystemEnvironmentVariables MinerSystemEnvironmentVariables { get; set; } = new MinerSystemEnvironmentVariables
+        {
+            DefaultSystemEnvironmentVariables = new Dictionary<string, string>()
+            {
+                {"GPU_MAX_ALLOC_PERCENT", "100"},
+                {"GPU_USE_SYNC_OBJECTS", "1"},
+                {"GPU_SINGLE_ALLOC_PERCENT", "100"},
+                {"GPU_MAX_HEAP_SIZE", "100"},
+                {"GPU_FORCE_64BIT_PTR", "1"}
+            },
+        };
+    }
+}
